backend/requirements.txt
Django>=4.2
djangorestframework
requests
xarray
cfgrib
eccodes
numpy
Pillow
beautifulsoup4
django-cors-headers
gunicorn